requirements.txt
actionlib==1.13.2
aiohttp==3.7.4.post0
angles==1.9.13
async-timeout==3.0.1
attrs==20.3.0
betse==1.1.1
bondpy==1.8.6
cachetools==4.2.1
camera-calibration==1.15.2
camera-calibration-parsers==1.12.0
catkin==0.8.9
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
controller-manager==0.19.4
controller-manager-msgs==0.19.4
cv-bridge==1.15.0
cycler==0.10.0
decorator==5.0.6
diagnostic-analysis==1.10.3
diagnostic-common-diagnostics==1.10.3
diagnostic-updater==1.10.3
dill==0.3.3
dynamic-reconfigure==1.7.1
fsspec==0.9.0
gazebo-plugins==2.9.1
gazebo-ros==2.9.1
gcsfs==0.8.0
gencpp==0.6.5
geneus==3.0.0
genlisp==0.4.18
genmsg==0.5.16
gennodejs==2.0.2
genpy==0.6.14
google-api-core==1.26.3
google-auth==1.28.1
google-auth-oauthlib==0.4.4
google-cloud-core==1.6.0
google-cloud-storage==1.37.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
idna==2.10
image-geometry==1.15.0
interactive-markers==1.12.0
joint-state-publisher==1.15.0
joint-state-publisher-gui==1.15.0
kiwisolver==1.3.1
laser-geometry==1.6.5
matplotlib==3.4.1
message-filters==1.15.9
multidict==5.1.0
numpy==1.20.2
oauthlib==3.1.0
packaging==20.9
pandas==1.2.3
Pillow==8.2.0
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-qt-binding==0.4.3
pytz==2021.1
PyYAML==5.4.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests==2.25.1
requests-oauthlib==1.3.0
resource-retriever==1.12.6
rosbag==1.15.9
rosboost-cfg==1.15.7
rosclean==1.15.7
roscreate==1.15.7
rosgraph==1.15.9
roslaunch==1.15.9
roslib==1.15.7
roslint==0.12.0
roslz4==1.15.9
rosmake==1.15.7
rosmaster==1.15.9
rosmsg==1.15.9
rosnode==1.15.9
rosparam==1.15.9
rospy==1.15.9
rosserial-arduino==0.9.1
rosserial-client==0.9.1
rosserial-python==0.9.1
rosservice==1.15.9
rostest==1.15.9
rostopic==1.15.9
rosunit==1.15.7
roswtf==1.15.9
rqt-action==0.4.9
rqt-bag==0.5.0
rqt-bag-plugins==0.5.0
rqt-console==0.4.11
rqt-dep==0.4.10
rqt-graph==0.4.14
rqt-gui==0.5.2
rqt-gui-py==0.5.2
rqt-image-view==0.4.16
rqt-launch==0.4.9
rqt-logger-level==0.4.11
rqt-moveit==0.5.9
rqt-msg==0.4.9
rqt-nav-view==0.5.7
rqt-plot==0.4.12
rqt-pose-view==0.5.10
rqt-publisher==0.4.9
rqt-py-common==0.5.2
rqt-py-console==0.4.9
rqt-reconfigure==0.5.3
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.13
rqt-robot-steering==0.5.12
rqt-runtime-monitor==0.5.8
rqt-rviz==0.6.1
rqt-service-caller==0.4.9
rqt-shell==0.4.10
rqt-srv==0.4.8
rqt-tf-tree==0.6.2
rqt-top==0.4.9
rqt-topic==0.4.12
rqt-web==0.4.9
rsa==4.7.2
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
rviz==1.14.4
scipy==1.6.2
seaborn==0.11.1
sensor-msgs==1.13.0
six==1.15.0
smach==2.5.0
smach-ros==2.5.0
smclib==1.8.6
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.5
tf2-kdl==0.7.5
tf2-py==0.7.5
tf2-ros==0.7.5
topic-tools==1.15.9
torch==1.8.1
typing-extensions==3.7.4.3
ujson==4.0.2
urllib3==1.26.4
xacro==1.14.5
yarl==1.6.3
